+++ src/lib/components/Modal.svelte
@@ -22,6 +22,6 @@
       <slot name="foot" {close} />
     </footer>
   </div>
-  <button type="button" class="du-modal-backdrop" aria-label="Close modal" on:click={() => close()}
+  <button type="button" class="du-modal-backdrop" aria-label="Close modal" onclick={() => close()}
   ></button>
 </div>