+++ src/routes/(app)/more/about/+page.svelte
@@ -69,7 +69,7 @@
           <ul>
             {#each links as link}
               <li>
-                <a href={link.href} on:click|preventDefault={(_e) => externalLink(link.href)}>
+                <a href={link.href} onclick={(e) => { e.preventDefault(); externalLink(link.href); }}>
                   <span class="icon is-small">
                     <i class={link.icon}></i>
                   </span>